Created wallet info card, added the UI tests and repository tests

+++ app/src/main/java/com/kevalpatel2106/blockchain/info/repository/dto/TransactionDto.kt
@@ -5,29 +5,29 @@ import com.kevalpatel2106.blockchain.info.bin.Transaction
 
 data class TransactionDto(
 
-    @field:SerializedName("fee")
-    var fee: Int,
+        @field:SerializedName("fee")
+        var fee: Long,
 
-    @field:SerializedName("result")
-    var result: Int,
+        @field:SerializedName("result")
+        var result: Long,
 
-    @field:SerializedName("balance")
-    var balance: Int,
+        @field:SerializedName("balance")
+        var balance: Long,
 
-    @field:SerializedName("time")
-    var time: Long,
+        @field:SerializedName("time")
+        var time: Long,
 
-    @field:SerializedName("hash")
-    var hash: String
+        @field:SerializedName("hash")
+        var hash: String
 ) {
 
     fun toEntity(): Transaction {
         return Transaction(
-            fee = fee,
-            result = result,
-            balance = balance,
-            time = time,
-            hash = hash
+                fee = fee,
+                result = result,
+                balance = balance,
+                time = time,
+                hash = hash
         )
     }
 }
